pyproject.toml
[project]
name = "modal-armor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.9.0",
    "black>=23.0.0",
    "bleach>=6.0.0",
    "chromadb>=0.4.0",
    "fastapi>=0.104.0",
    "flake8>=6.1.0",
    "google-genai>=1.0.0",
    "guardrails-ai>=0.6.0",
    "httpx>=0.25.0",
    "loguru>=0.7.0",
    "mypy>=1.5.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pre-commit>=3.4.0",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "prometheus-client>=0.19.0",
    "pydantic>=2.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0.0",
    "redis>=5.0.0",
    "requests>=2.31.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",
    "slowapi>=0.1.9",
    "spacy>=3.7.0",
    "sqlalchemy>=2.0.0",
    "toml>=0.10.2",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "uvicorn[standard]>=0.24.0",
    "yara-python>=4.3.0",
]
